PR review - moves to controllers pkg and uses log.WithValues
