+++ api/v1alpha1/watermarkpodautoscaler_types.go
@@ -6,7 +6,6 @@
 package v1alpha1
 
 import (
-	"encoding/json"
 	autoscalingv2 "k8s.io/api/autoscaling/v2beta1"
 	v1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/resource"
@@ -204,31 +203,4 @@ type WatermarkPodAutoscalerList struct {
 
 func init() {
 	SchemeBuilder.Register(&WatermarkPodAutoscaler{}, &WatermarkPodAutoscalerList{})
-}
-
-// GetLogAttrs takes a list of log attrs that would be sent to logger.info, and adds user specified key value attribute
-// pairs from the ad.datadoghq.com/attributes annotation of the WPA spec.
-// logAttrs should be pairs of key/values, with each key being a string.
-// This func returns a slice! To pass this as variadic parameters to logger.info, you need the ellipses, so use wpa.GetLogAttrs()...
-func (wpa *WatermarkPodAutoscaler) GetLogAttrs(logAttrs ...interface{}) []interface{} {
-	var logAttributes []interface{}
-
-	for _, v := range logAttrs {
-		logAttributes = append(logAttributes, v)
-	}
-
-	customAttrsStr := wpa.ObjectMeta.Annotations["ad.datadoghq.com/attributes"]
-	var customAttrs map[string]interface{}
-	err := json.Unmarshal([]byte(customAttrsStr), &customAttrs)
-	if err != nil {
-		// Someone put invalid JSON in their annotation, don't want to spam controller logs with errors for that.
-		// Just continue with the log attributes defined by the controller.
-		return logAttributes
-	}
-
-	for k,v := range customAttrs {
-		logAttributes = append(logAttributes, k)
-		logAttributes = append(logAttributes, v)
-	}
-	return logAttributes
 }
